+++ doc/Changelog.md
@@ -3,6 +3,10 @@ Changelog {#Changelog}
 
 # Release 1.13 (git master)
 
+* [585](https://github.com/Eyescale/Equalizer/pull/585):
+  Deflect proxy: update to Deflect 0.12 v5 API, use new PAN event to move
+  camera instead of switching between pan/rotate with TAP_AND_HOLD event,
+  handle new pinch event to zoom.
 * [584](https://github.com/Eyescale/Equalizer/pull/584):
   Removed broken CUDA/ComputeContext support
 * [578](https://github.com/Eyescale/Equalizer/pull/578):